VBA emailing module
This is to enable auto-emailing in Excel. Can be used to attach multiple attachments for each client, like reports & invoices.
Column E is the identifiers of client code, while column V specifies their email addresses. The names of pdf attachments all contain client code in the start, so the program can pick up the attachments according to their code.